Add stickman in "gallows"

+++ src/gallows.py
@@ -4,6 +4,72 @@ from .database import database
 
 from uuid import uuid4
 
+stickman = [
+"""
+___________
+|         |
+|         
+|        
+|        
+|
+|
+""",
+"""
+___________
+|         |
+|         0
+|        
+|        
+|
+|
+""",
+"""
+___________
+|         |
+|         0
+|         |
+|        
+|
+|
+""",
+"""
+___________
+|         |
+|         0
+|        /|
+|        
+|
+|
+""",
+r"""
+___________
+|         |
+|         0
+|        /|\
+|        
+|
+|
+""",
+r"""
+___________
+|         |
+|         0
+|        /|\
+|        / 
+|
+|
+""",
+r"""
+___________
+|         |
+|         0
+|        /|\
+|        / \
+|
+|
+"""
+]
+
 
 def gallows_suggestion(text, chat_id):
     game = database.gallows.find_one({'chat': chat_id})
@@ -13,14 +79,21 @@ def gallows_suggestion(text, chat_id):
         for i in range(len(splitted_word)):
             if splitted_word[i] == text:
                 has_letter = True
-                game['word_in_underlines'][i] = text
         if has_letter:
+            if text in game['word_in_underlines']:
+                bot.send_message(chat_id, 'Эта буква уже угадана.')
+                return
+            for i in range(len(splitted_word)):
+                if splitted_word[i] == text:
+                    game['word_in_underlines'][i] = text
             if game['word_in_underlines'] == splitted_word:
-                database.gallows.delete_one({'_id': game['_id']})
                 bot.send_message(
                     chat_id,
-                    f'{"".join(game["word_in_underlines"])}\nВы победили'
+                    f'<code>{stickman[-(game["attempts"])]}</code>\n'
+                    f'Победа.\nСлово: {" ".join(game["word_in_underlines"])}',
+                    parse_mode='HTML'
                 )
+                database.gallows.delete_one({'_id': game['_id']})
                 return
             database.gallows.update_one(
                 {'chat': game['chat']},
@@ -28,14 +101,26 @@ def gallows_suggestion(text, chat_id):
             )
             bot.send_message(
                 chat_id,
-                f'{"".join(game["word_in_underlines"])}'
+                f'<code>{stickman[-(game["attempts"])]}</code>\n'
+                f'Слово: {" ".join(game["word_in_underlines"])}',
+                parse_mode='HTML'
             )
             return
         database.gallows.update_one(
             {'chat': game['chat']},
             {'$inc': {'attempts': -1}}
         )
+        if game['attempts']-1 > 1:
+            bot.send_message(
+                chat_id, 
+                f'<code>{stickman[-(game["attempts"]-1)]}</code>'
+                f'\nСлово: {" ".join(game["word_in_underlines"])}',
+                parse_mode='HTML'
+            )
+            return
         bot.send_message(
-            chat_id, 
-            f'{game["word_in_underlines"]}\nОсталось попыток: {game["attempts"] - 1}'
+            chat_id,
+            f'<code>{stickman[-(game["attempts"]-1)]}</code>\n'
+            f'Вы проиграли. Загаданным словом было: {game["word"]}.'
         )
+        database.gallows.delete_one({'_id': game['_id']})
